split Document/utils into shared/ui
