+++ packages/fields-document/src/DocumentEditor/utils.ts
@@ -1,21 +1,17 @@
-'use client'
-import React, { useCallback, useEffect, useRef, useState, useContext } from 'react'
 import {
-  Editor,
-  Node,
-  type NodeEntry,
-  Path,
-  Transforms,
-  Element,
-  type PathRef,
-  Text,
   type Location,
+  type NodeEntry,
+  type PathRef,
   type Point,
+  Editor,
+  Element,
+  Node,
+  Path,
+  Text,
+  Transforms,
 } from 'slate'
 import { ReactEditor } from 'slate-react'
 import { type ElementFromValidation } from '../structure-validation'
-
-export { useSlateStatic as useStaticEditor } from 'slate-react'
 
 export type Mark =
   | 'bold'
@@ -38,10 +34,15 @@ export const allMarks: Mark[] = [
   'keyboard',
 ]
 
-export const isElementActive = (
+const IS_MAC =
+  typeof window != 'undefined' && /Mac|iPod|iPhone|iPad/.test(window.navigator.platform)
+
+export const modifierKeyText = IS_MAC ? '⌘' : 'Ctrl'
+
+export function isElementActive (
   editor: Editor,
   format: Exclude<ElementFromValidation, { text: string }>['type']
-) => {
+) {
   const [match] = Editor.nodes(editor, {
     match: n => n.type === format,
   })
@@ -72,66 +73,6 @@ export function moveChildren (
       Transforms.moveNodes(editor, { at: childPath, to })
     }
   }
-}
-
-// this ensures that when changes happen, they are immediately shown
-// this stops the problem of a cursor resetting to the end when a change is made
-// because the changes are applied asynchronously
-export function useElementWithSetNodes<TElement extends Element> (
-  editor: Editor,
-  element: TElement
-) {
-  const [state, setState] = useState({ element, elementWithChanges: element })
-  if (state.element !== element) {
-    setState({ element, elementWithChanges: element })
-  }
-
-  const elementRef = useRef(element)
-
-  useEffect(() => {
-    elementRef.current = element
-  })
-
-  const setNodes = useCallback(
-    (changesOrCallback: Partial<TElement> | ((current: TElement) => Partial<TElement>)) => {
-      const currentElement = elementRef.current
-      const changes =
-        typeof changesOrCallback === 'function'
-          ? changesOrCallback(currentElement)
-          : changesOrCallback
-      Transforms.setNodes(editor, changes, { at: ReactEditor.findPath(editor, currentElement) })
-      setState({
-        element: currentElement,
-        elementWithChanges: { ...currentElement, ...changes },
-      })
-    },
-    [editor]
-  )
-  return [state.elementWithChanges, setNodes] as const
-}
-
-export function useEventCallback<Func extends (...args: any) => any>(callback: Func): Func {
-  const callbackRef = useRef(callback)
-  const cb = useCallback((...args: any[]) => {
-    return callbackRef.current(...args)
-  }, [])
-  useEffect(() => {
-    callbackRef.current = callback
-  })
-  return cb as any
-}
-
-const IS_MAC =
-  typeof window != 'undefined' && /Mac|iPod|iPhone|iPad/.test(window.navigator.platform)
-
-export const modifierKeyText = IS_MAC ? '⌘' : 'Ctrl'
-
-const ForceValidationContext = React.createContext(false)
-
-export const ForceValidationProvider = ForceValidationContext.Provider
-
-export function useForceValidation () {
-  return useContext(ForceValidationContext)
 }
 
 export function insertNodesButReplaceIfSelectionIsAtEmptyParagraphOrHeading (
